+++ irc3/plugins/ctcp.py
@@ -10,7 +10,7 @@ __doc__ = '''
 
 Usage::
 
-    >>> bot = IrcBot(ctcp=dict(foo='bar'))
+    >>> bot = IrcBot(ctcp=dict(foo='bar'), ctcp_max_replies=5)
     >>> bot.include('irc3.plugins.ctcp')
 
 Try to send a ``CTCP FOO``::
@@ -20,14 +20,50 @@ Try to send a ``CTCP FOO``::
     >>> print(bot.sent[0].replace('\x01', '01'))
     NOTICE gawel :01FOO bar01
 '''
-from irc3 import event
-from irc3 import rfc
+from datetime import datetime
+from irc3.compat import Queue
+from irc3.compat import QueueFull
+import irc3
 
 
-@event(rfc.CTCP)
-def ctcp(bot, mask=None, target=None, ctcp=None, **kw):
+@irc3.plugin
+class CTCP(object):
     """ctcp replies"""
-    lctcp = ctcp.lower()
-    if lctcp in bot.config.ctcp:
-        data = bot.config.ctcp[lctcp].format(**bot.config)
-        bot.ctcp_reply(mask.nick, '%s %s' % (ctcp, data))
+
+    def __init__(self, bot):
+        maxsize = int(bot.config.get('ctcp_max_replies', 5))
+        self.queue = Queue(loop=bot.loop, maxsize=maxsize)
+        self.handle = None
+        self.event = irc3.event(irc3.rfc.CTCP, self.on_ctcp)
+        bot.attach_events(self.event)
+        self.bot = bot
+
+    def send_replies(self):
+        if self.queue.full():  # pragma: no cover
+            # ignore requests when the queue is full / flood
+            self.queue = Queue(loop=self.bot.loop)
+        else:
+            while not self.queue.empty():
+                target, ctcp = self.queue.get_nowait()
+                data = self.bot.config.ctcp[ctcp].format(now=datetime.now(),
+                                                         **self.bot.config)
+                self.bot.ctcp_reply(target, '%s %s' % (ctcp.upper(), data))
+        self.handle = None
+
+    def on_ctcp(self, mask=None, target=None, ctcp=None, **kw):
+        lctcp = ctcp.lower()
+        if lctcp in self.bot.config.ctcp:
+            try:
+                self.queue.put_nowait((mask.nick, lctcp))
+            except QueueFull:  # pragma: no cover
+                self.bot.log.warn('CTCP Flood detected. '
+                                  'Ignoring requests for 30s')
+                self.bot.detach_events(self.event)
+                self.bot.loop.call_later(
+                    30, self.bot.attach_events, self.event)
+            else:
+                if self.bot.config.testing:
+                    self.send_replies()
+                elif self.handle is None:  # pragma: no cover
+                    self.handle = self.bot.loop.call_later(
+                        2, self.send_replies)